added io pkg; refactor input output functionality to io pkg
