+++ commands/add.go
@@ -6,15 +6,16 @@ import (
 
 	"github.com/nurali-techie/rc/cli"
 	"github.com/nurali-techie/rc/domain"
+	"github.com/nurali-techie/rc/io"
 	"github.com/nurali-techie/rc/service"
 )
 
 type addCommand struct {
 	service service.CommandService
-	input   cli.Input
+	input   io.Input
 }
 
-func NewAddCommand(service service.CommandService, input cli.Input) cli.Command {
+func NewAddCommand(service service.CommandService, input io.Input) cli.Command {
 	addCmd := new(addCommand)
 	addCmd.service = service
 	addCmd.input = input
