+++ commands/get.go
@@ -5,15 +5,16 @@ import (
 	"fmt"
 
 	"github.com/nurali-techie/rc/cli"
+	"github.com/nurali-techie/rc/io"
 	"github.com/nurali-techie/rc/service"
 )
 
 type getCommand struct {
 	service service.CommandService
-	output  cli.Output
+	output  io.Output
 }
 
-func NewGetCommand(service service.CommandService, output cli.Output) cli.Command {
+func NewGetCommand(service service.CommandService, output io.Output) cli.Command {
 	getCmd := new(getCommand)
 	getCmd.service = service
 	getCmd.output = output
