+++ io/input.go
@@ -1,4 +1,4 @@
-package cli
+package io
 
 type Input interface {
 	GetContent() (string, error)
