+++ io/output.go
@@ -1,4 +1,4 @@
-package cli
+package io
 
 type Output interface {
 	SetContent(content string) error
